+++ app/src/main/java/com/vktrsansara/app/caveviewer/presentation/settings/CursorColorPickerDialog.kt
@@ -81,7 +81,7 @@ fun CursorColorPickerDialog(
                 modifier = Modifier.fillMaxWidth()
             ) {
                 items(CursorPaletteColors) { colorLong ->
-                    val color = Color(colorLong)
+                    val color = Color(colorLong.toInt())
                     val isSelected = colorLong == selectedColor
 
                     Box(
